Run test container as root
Local directory belongs to UID 1001, while the container's normal user has UID 1654 (intentionally set by .NET SDK's Dockerfile).
The simplest way to be able to write to the bind-mounted directory in the container is to be root.
This is also, apparently, what GitLab does. (GitLab runners also seem to run as root locally.)